+++ Development/Source/script_interpreter_func_fe3d_collision.cpp
@@ -8,7 +8,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		auto types = { ScriptValueType::DECIMAL, ScriptValueType::DECIMAL }; // CameraHeight + cameraSpeed
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			_fe3d.collision_enableCameraTerrainResponse(arguments[0].getDecimal(), arguments[1].getDecimal());
 			returnValues.push_back(ScriptValue(_fe3d, ScriptValueType::EMPTY));
@@ -42,7 +42,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		};
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			_fe3d.collision_setCameraBoxSize(
 				arguments[0].getDecimal(), arguments[1].getDecimal(), arguments[2].getDecimal(),
@@ -59,7 +59,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		};
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			_fe3d.collision_enableCameraResponse(arguments[0].getBoolean(), arguments[1].getBoolean(), arguments[2].getBoolean());
 			returnValues.push_back(ScriptValue(_fe3d, ScriptValueType::EMPTY));
@@ -79,7 +79,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		auto types = { ScriptValueType::STRING, ScriptValueType::STRING, ScriptValueType::STRING }; // ModelEntity ID + aabbEntity part ID + direction
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Validate direction
 			if (arguments[2].getString() != "X" && arguments[2].getString() != "Y" && 
@@ -135,7 +135,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		auto types = { ScriptValueType::STRING }; // Direction
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Validate direction
 			if (arguments[0].getString() != "X" && arguments[0].getString() != "Y" &&
@@ -174,7 +174,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		auto types = { ScriptValueType::STRING, ScriptValueType::STRING }; // aabbEntity ID + Direction
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Validate direction
 			if (arguments[1].getString() != "X" && arguments[1].getString() != "Y" && 
@@ -220,7 +220,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		auto types = { ScriptValueType::STRING }; // Direction
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Validate direction
 			if (arguments[0].getString() != "X" && arguments[0].getString() != "Y" &&
@@ -263,7 +263,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		};
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Temporary values
 			string result = "";
@@ -354,7 +354,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		};
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Temporary values
 			string result = "";
@@ -409,7 +409,7 @@ bool ScriptInterpreter::_executeFe3dCollisionFunction(const string& functionName
 		};
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Temporary values
 			string result = "";
