+++ Development/Source/script_interpreter_func_fe3d_terrain.cpp
@@ -20,7 +20,7 @@ bool ScriptInterpreter::_executeFe3dTerrainEntityFunction(const string& function
 		auto types = { ScriptValueType::DECIMAL, ScriptValueType::DECIMAL };
 
 		// Validate arguments
-		if (_validateListValueAmount(arguments, types.size()) && _validateListValueTypes(arguments, types))
+		if (_validateListValueAmount(arguments, static_cast<unsigned int>(types.size())) && _validateListValueTypes(arguments, types))
 		{
 			// Validate terrain existence
 			if (_validateFe3dTerrainEntity())
